Move to sources root.
It was done due to the fact that an @EnableAutoConfiguration, annotated via @SpringBootApplication, has to be in the root of the sources. From docs: "It is generally recommended that you place @EnableAutoConfiguration [...] in a root package so that all sub-packages and classes can be searched." Now database's tables are auto-created properly.